+++ xbmc/cores/paplayer/PAPlayer.h
@@ -126,9 +126,9 @@ private:
   void SoftStart(bool wait = false);
   void SoftStop(bool wait = false, bool close = true);
   void CloseAllStreams(bool fade = true);
-  void ProcessStreams(float &delay, float &buffer);
+  void ProcessStreams(double &delay, double &buffer);
   bool PrepareStream(StreamInfo *si);
-  bool ProcessStream(StreamInfo *si, float &delay, float &buffer);
+  bool ProcessStream(StreamInfo *si, double &delay, double &buffer);
   bool QueueData(StreamInfo *si);
   __int64 GetTotalTime64();
 };
